float B_ul and B_lu

+++ globals.h
@@ -204,8 +204,8 @@ struct TransitionLines {
   MPI_shared_array<const int> ionindex;  // The same for the elements ion
   MPI_shared_array<const int> uniquelevelindex_lower;  // globally unique index of the lower level of the transition
   MPI_shared_array<const int> uniquelevelindex_upper;  // globally unique index of the upper level of the transition
-  MPI_shared_array<const double> B_ul;  // Einstein B coefficient for stimulated emission
-  MPI_shared_array<const double> B_lu;  // Einstein B coefficient for absorption (B_lu = (g_u/g_l) * B_ul)
+  MPI_shared_array<const float> B_ul;  // Einstein B coefficient for stimulated emission
+  MPI_shared_array<const float> B_lu;  // Einstein B coefficient for absorption (B_lu = (g_u/g_l) * B_ul)
 };
 inline TransitionLines linelist{};
 inline int nlines{-1};
